NinelTour_8/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
future==0.18.2
psycopg2==2.9.1
pytz==2021.3
serializers==0.2.4
sqlparse==0.4.2
